Make Frizzle objects that only hold weak references to documents.
The existing workarounds of making the Frizzle cache hold weak values
or soft values is not sufficient to solve the problem:

1. Holding weak values essentially means that the cache entries fall
   out as soon as the `Frizzle` object reference is released. This
   causes lots of extra reloads.

2. Holding soft values, while `Frizzle` objects hold a strong reference
   to `Document` objects, means that `Document`s' lifetimes are extended
   to the corresponding `Frizzle`s' lifetimes, and essentially makes a
   soft-keys cache. You do not want to use soft-keys caches for any
   reason: http://stackoverflow.com/a/7618134

Really, the real problem is the dependence of the cache key's lifetime
on that of the cache value, due to the latter possessing a strong
reference to the former. By weakening said reference, the cache key's
lifetime becomes independent of the cache value's lifetime.